+++ CODE/GAMEPLAY/COMPONENT/GAMEPLAY_COMPONENT_MANAGER.cpp
@@ -24,7 +24,7 @@ GAMEPLAY_COMPONENT_MANAGER::GAMEPLAY_COMPONENT_MANAGER() :
 
 GAMEPLAY_COMPONENT_MANAGER::~GAMEPLAY_COMPONENT_MANAGER() {
 
-    for ( int i = 0; i < InternalVector.size(); i++ ) {
+    for ( size_t i = 0; i < InternalVector.size(); i++ ) {
         
         CORE_MEMORY_ALLOCATOR_Free( InternalVector[ i ].MemoryArray );
     }
@@ -61,7 +61,7 @@ void GAMEPLAY_COMPONENT_MANAGER::SaveToStream( CORE_DATA_STREAM & stream ) {
         
         stream << InternalVector.size();
         
-        for ( int i = 0; i< InternalVector.size(); i++ ) {
+        for ( size_t i = 0; i< InternalVector.size(); i++ ) {
             
             stream.InputBytes((uint8_t *) &InternalVector[ i ].MemoryArray, sizeof(GAMEPLAY_COMPONENT_ENTITY) * GAMEPLAY_COMPONENT_BASE_COUNT );
             
@@ -92,7 +92,7 @@ void GAMEPLAY_COMPONENT_MANAGER::LoadFromStream( CORE_DATA_STREAM & stream ) {
     
     InternalVector.resize( size );
     
-    for ( int i = 0; i< size; i++ ) {
+    for ( size_t i = 0; i< size; i++ ) {
         
         size_t b = sizeof(GAMEPLAY_COMPONENT_ENTITY) * GAMEPLAY_COMPONENT_BASE_COUNT;
         stream.OutputBytes((uint8_t *) InternalVector[ i ].MemoryArray, b );
