+++ CODE/GAMEPLAY/COMPONENT/GAMEPLAY_COMPONENT_POSITION.cpp
@@ -72,7 +72,7 @@ void GAMEPLAY_COMPONENT_POSITION::SaveToStream( CORE_DATA_STREAM & stream ) {
         
         stream << InternalVector->size();
         
-        for ( int i = 0; i< InternalVector->size(); i++ ) {
+        for ( size_t i = 0; i< InternalVector->size(); i++ ) {
             
             stream.InputBytes((uint8_t *) (*InternalVector)[ i ].MemoryArray, sizeof(GAMEPLAY_COMPONENT_POSITION) * GAMEPLAY_COMPONENT_BASE_COUNT );
             
@@ -89,7 +89,7 @@ void GAMEPLAY_COMPONENT_POSITION::LoadFromStream( CORE_DATA_STREAM & stream ) {
     
     InternalVector->resize( size );
     
-    for ( int i = 0; i< size; i++ ) {
+    for ( size_t i = 0; i< size; i++ ) {
         
         InitializeMemory<INTERNAL_ARRAY, GAMEPLAY_COMPONENT_POSITION>( *InternalVector, i );
