+++ CODE/GAMEPLAY/COMPONENT/GAMEPLAY_COMPONENT_RENDER.cpp
@@ -100,7 +100,7 @@ void GAMEPLAY_COMPONENT_RENDER::SaveToStream( CORE_DATA_STREAM & stream ) {
         
         stream << InternalVector->size();
         
-        for ( int i = 0; i< InternalVector->size(); i++ ) {
+        for ( size_t i = 0; i< InternalVector->size(); i++ ) {
             
             stream.InputBytes((uint8_t *) (*InternalVector)[ i ].MemoryArray, sizeof(GAMEPLAY_COMPONENT_RENDER) * GAMEPLAY_COMPONENT_BASE_COUNT );
             
@@ -117,7 +117,7 @@ void GAMEPLAY_COMPONENT_RENDER::LoadFromStream( CORE_DATA_STREAM & stream ) {
     
     InternalVector->resize( size );
     
-    for ( int i = 0; i< size; i++ ) {
+    for ( size_t i = 0; i < size; i++ ) {
         
         InitializeMemory<INTERNAL_ARRAY_R, GAMEPLAY_COMPONENT_RENDER>( *InternalVector, i );
